rm old get up and go and mock uploads tab
