+++ Utah/Supporting Files/Localizable.strings
@@ -7,14 +7,6 @@
 //
 
 
-// MARK: - Contacts
-"CONTACTS_TAB_TITLE" = "Contacts";
-
-
-// MARK: - Mock Upload Data Storage Provider
-"MOCK_UPLOAD_TAB_TITLE" = "Mock Upload";
-
-
 // MARK: - Schedule
 "SCHEDULE_TAB_TITLE" = "Questions";
