--- UtahModules/Sources/UtahMockDataStorageProvider/Resources/en.lproj/Localizable.strings
@@ -1,13 +1,0 @@
-//
-// This source file is part of the CS342 2023 Utah Team Application project
-//
-// SPDX-FileCopyrightText: 2023 Stanford University
-//
-// SPDX-License-Identifier: MIT
-//
-
-// MARK: - Mock Upload Data Storage Provider
-"MOCK_UPLOAD_LIST_PLACEHOLDER" = "The Mock Upload Data Storage Provider will display all uploads that would be triggered by the FHIR standard.";
-"MOCK_UPLOAD_LIST_TITLE" = "Mock Upload";
-"MOCK_UPLOAD_DETAIL_HEADER" = "Upload Header";
-"MOCK_UPLOAD_DETAIL_BODY" = "Upload Body";
