+++ UtahModules/Sources/UtahSchedule/WIQ/WIQViewController.swift
@@ -6,6 +6,7 @@
 // SPDX-License-Identifier: MIT
 //
 
+// swiftlint:disable closure_body_length
 // swiftlint:disable line_length
 
 import ResearchKit
