+++ UtahUITests/TrendsTests.swift
@@ -32,10 +32,9 @@ class TrendsTests: XCTestCase {
         app.activate()
         sleep(5)
         let newVal = (prevValue ?? 0.0) + 42
-        try navigateToMockUpload()
-        try assertObservationCellPresent(true, pressIfPresent: true)
-        try assertObservationCellPresent(true, pressIfPresent: false)
-        
+        // Need to navigate to another tab first to refresh number
+        XCTAssertTrue(app.tabBars["Tab Bar"].buttons["Questions"].waitForExistence(timeout: 2))
+        app.tabBars["Tab Bar"].buttons["Questions"].tap()
         try navigateToTrends()
         XCTAssert(app.staticTexts[String(newVal)].waitForExistence(timeout: 2))
     }
@@ -48,26 +47,5 @@ class TrendsTests: XCTestCase {
         XCTAssertTrue(app.staticTexts["Daily Step Count"].waitForExistence(timeout: 2))
         XCTAssertTrue(app.staticTexts["Last EFS Survey Score"].waitForExistence(timeout: 2))
     }
-    private func navigateToMockUpload() throws {
-        let app = XCUIApplication()
-        
-        XCTAssertTrue(app.tabBars["Tab Bar"].buttons["Mock Upload"].waitForExistence(timeout: 2))
-        app.tabBars["Tab Bar"].buttons["Mock Upload"].tap()
-    }
     
-    private func assertObservationCellPresent(_ shouldBePresent: Bool, pressIfPresent: Bool = true) throws {
-        let app = XCUIApplication()
-        
-        let observationText = "/Observation/"
-        let predicate = NSPredicate(format: "label CONTAINS[c] %@", observationText)
-        
-        if shouldBePresent {
-            XCTAssertTrue(app.staticTexts.containing(predicate).firstMatch.waitForExistence(timeout: 2))
-            if pressIfPresent {
-                app.staticTexts.containing(predicate).firstMatch.tap()
-            }
-        } else {
-            XCTAssertFalse(app.staticTexts.containing(predicate).firstMatch.waitForExistence(timeout: 2))
-        }
-    }
 }
