+++ UtahUITests/WalkTestTests.swift
@@ -1,0 +1,35 @@
+//
+// This source file is part of the CS342 2023 Utah Team Application project
+//
+// SPDX-FileCopyrightText: 2023 Stanford University
+//
+// SPDX-License-Identifier: MIT
+//
+
+import XCTest
+
+
+class WalkTestTests: XCTestCase {
+    override func setUpWithError() throws {
+        try super.setUpWithError()
+        
+        continueAfterFailure = false
+        
+        let app = XCUIApplication()
+        app.launchArguments = ["--skipOnboarding"]
+        app.launch()
+    }
+    
+    func testWalkTest() throws {
+        let app = XCUIApplication()
+        
+        XCTAssertTrue(app.tabBars["Tab Bar"].buttons["Questions"].waitForExistence(timeout: 2))
+        app.tabBars["Tab Bar"].buttons["Questions"].tap()
+
+        XCTAssertTrue(app.buttons["6 Minute Walk Test (active task)"].waitForExistence(timeout: 2))
+        app.buttons["6 Minute Walk Test (active task)"].tap()
+        XCTAssertTrue(app.staticTexts["Fitness"].waitForExistence(timeout: 2))
+        XCTAssertTrue(app.buttons["Next"].waitForExistence(timeout: 2))
+        app.buttons["Next"].tap()
+    }
+}
